Venice integ test improvements (#8350)
integ test fixes

more venice integ improvements

add a readme

venice integ fixes

more fixes

Fix sanity failures

Fix test coverage

+++ test/utils/elastic.go
@@ -153,8 +153,10 @@ func StopElasticsearch(name, authDir string) error {
 
 	log.Info("stopping elasticsearch ...")
 
-	defer certs.DeleteTLSCredentials(authDir)
-	defer os.RemoveAll(authDir)
+	if authDir != "" {
+		defer certs.DeleteTLSCredentials(authDir)
+		defer os.RemoveAll(authDir)
+	}
 
 	cmd := []string{"rm", "-f", name}
